+++ R/theme_zzz.R
@@ -1,5 +1,5 @@
 theme_dictionary <- list(
-  "default" = theme_default,
+  "default" = theme_tinytable,
   "grid" = theme_grid,
   "revealjs" = theme_revealjs,
   "rotate" = theme_rotate,